feat: extract main logic in a script and use the rest as an extension
