+++ examples/hugo/operations.graphql
@@ -8,6 +8,8 @@ query Generate($src: FSID!) {
 
 query Deploy($contents: FSID!, $bucketURL: String!, $subdir: String, $token: SecretID!) {
   hugo {
-    deploy(contents: $contents, bucketURL: $bucketURL, subdir: $subdir, token: $token) 
+    deploy(contents: $contents, bucketURL: $bucketURL, subdir: $subdir, token: $token) {
+	    id
+    }
   }
 }
